Modified the listings Controller
Our Listings Controller's create function that handles the creation of listings now accepts an optional parameter called $category that gets the default value of null if not specified by our view. It works with  the validCategory function within our controller to validate whether or not a valid category was provided from the view. If valid then it proceeds and shows the  create listing view providing it with the appropriate category.