Worked around the Create Listing view.
Reworked the values within the work as they were badly written due to copy pasting. Added a hidden value pasting using php the category in that we got from the previous view. Also, a new error section ontop of the page that displays all errors (if any) that were forwarded by our countroller.