Server Sided Validation of the data provided by the Create Listing view.
Basic validation of the create listing view. Has a couple of notable flaws. The Country for instance is a set of strings but I couldn't be bothered. 

We haven't set up a view yet for displaying the listings so it just dumps all the data from the request. For debug purposes of couse.