Changes to Migrations; Create Listing View and Listing Controller.
The $ee for the listing is passed dynamically to the view. Added two values for Listings in the migrations. One called featued. It's boolean, if true then this listing is featured. This will be toggled by admins to randomly display this listing at the home page. And one for brand where the user can enter their device's brand.